refactor: 테이블 스크롤 상태 뱃지부분 보여주기
- 반응형 변경 시점에서 스크롤 초기값을 맨 끝으로 만든 것
